move the json building to the consumer
this should save some memory for the queue since it's only storing the
timestamp and message now and not all of the other stuff for every
message.

for 100k queue entries before this change I was running about 104MB RSS.
after this change I'm running about 96MB RSS. Not much, but there's
probably plenty of room for optimization later!