multiprocessing :(
so, with multiprocessing, if I killed both of the processes at once it
would break the pipe between them, so I'm moving back to threading and
gonna just make that work. It works as is, but now I need to rethink my
child thread as if it breaks it's going to end the world.